+++ core/moengage_flutter_platform_interface/lib/src/moengage_flutter_platform_interface.dart
@@ -1,16 +1,15 @@
 import 'package:plugin_platform_interface/plugin_platform_interface.dart';
 
-import '../moengage_flutter_platform_interface.dart';
 import '../src/model/app_status.dart';
 import '../src/model/gender.dart';
 import '../src/model/geo_location.dart';
+import '../src/model/inapp/self_handled_campaigns_data.dart';
 import '../src/model/moe_init_config.dart';
 import '../src/model/permission_type.dart';
 import '../src/model/properties.dart';
 import '../src/model/push/moe_push_service.dart';
 import 'internal/method_channel_moengage_flutter.dart';
 import 'model/inapp/nudge_position.dart';
-import 'model/inapp/self_handled_data.dart';
 import 'model/user_deletion_data.dart';
 
 /// Platform Interface for MoEngage Flutter Plugin
@@ -222,12 +221,13 @@ abstract class MoEngageFlutterPlatform extends PlatformInterface {
 
   /// Get Multiple Self Handled InApps
   /// @returns - Instance of [Future] of type [SelfHandledCampaignsData]
-  /// @since TODO : Add Version.
+  /// @since 3.1.0
   Future<SelfHandledCampaignsData> getSelfHandledInApps(String appId) =>
       throw UnimplementedError(
           'getSelfHandledInApps() not implemented for Platform');
 
   /// Register for Provisional Push
+  /// @since 3.1.0
   void registerForProvisionalPush() => throw UnimplementedError(
       'registerForProvisionalPush() not implemented for Platform');
 }