+++ cards/moengage_cards/CHANGELOG.md
@@ -1,4 +1,11 @@
 # MoEngage Cards Plugin
+
+# Next Release
+
+## Next Version
+- Android
+    - Bundling `cards-core` dependency with the plugin
+
 # 07-08-2024
 
 ## 4.0.0