+++ cards/moengage_cards/CHANGELOG.md
@@ -5,6 +5,8 @@
 ## Next Version
 - Android
     - Bundling `cards-core` dependency with the plugin
+    - Refresh cards on uniqueID set
+    - `cards-core` version updated to `2.3.0`
 
 # 07-08-2024
 
@@ -19,6 +21,16 @@
 - Exact version pinning for Cards Module dependencies
 - iOS
   - Updated MoEngageCards to `4.17.0`
+
+# Next Release date
+
+## Next version
+- Refresh cards on uniqueID set
+- Breaking Change: 
+
+|                     Then                      |            Now                     |
+|:---------------------------------------------:|:----------------------------------:|
+| MoEngageCards#setAppOpenCardsSyncListener()   | MoEngageCards#setSyncCompleteListener() |
 
 # 15-07-2024